tidied up code
