+++ src/frontend/views/plot/__init__.py
@@ -1,0 +1,1 @@
+from .plot_ import Bar, Network, Scatter
